refactor: delegate all routers to controllers
